+++ packages/web_socket_client/lib/src/connection.dart
@@ -1,0 +1,48 @@
+import 'dart:async';
+
+import 'package:ndk_web_socket_client/ndk_web_socket_client.dart';
+
+/// {@template connection_controller}
+/// A WebSocket connection controller.
+/// {@endtemplate}
+class ConnectionController extends Connection {
+  /// {@macro connection_controller}
+  ConnectionController()
+      : _state = const Connecting(),
+        _controller = StreamController<ConnectionState>.broadcast();
+
+  ConnectionState _state;
+  final StreamController<ConnectionState> _controller;
+
+  @override
+  ConnectionState get state => _state;
+
+  @override
+  StreamSubscription<ConnectionState> listen(
+    void Function(ConnectionState event)? onData, {
+    Function? onError,
+    void Function()? onDone,
+    bool? cancelOnError,
+  }) {
+    return _stream.distinct().listen(
+          onData,
+          onError: onError,
+          onDone: onDone,
+          cancelOnError: cancelOnError,
+        );
+  }
+
+  Stream<ConnectionState> get _stream async* {
+    yield _state;
+    yield* _controller.stream;
+  }
+
+  /// Notifies listeners of a new connection [state].
+  void add(ConnectionState state) {
+    _state = state;
+    _controller.add(state);
+  }
+
+  /// Closes the controller's stream.
+  void close() => _controller.close();
+}